Added code for uploading images
This code takes care of the functionality of taking pictures with the camera, uploading that picture to firebase. The pictures can be uploaded through the camera or gallery and the picture will be shown. Then, the user has the option to download the image, which will give the download url and the downloaded image to the user. Pretty sure that the downloaded image or the image taken from camera gets saved in the gallery.